requirements.txt
annotated-types
asn1crypto
certifi
cffi
charset-normalizer
cryptography
fastecdsa
idna
lxml
oscrypto
pillow
pyasice
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyivxv
pyzbar
requests
typing-inspection
typing_extensions
urllib3
